OK-376: Lisätty pistelaskurin näyttäminen suosikien vertailussa ja järjestelty vertailu-sivun koodia uudelleen
