+++ src/main/app/public/locales/fi/translation.json
@@ -871,6 +871,8 @@
     "valintakoe": "Koe- tai lisänäyttö",
     "urheilijan-amm-koulutus": "Urheilijan ammatillinen koulutus",
     "valitse-vertailtavat-tiedot": "Valitse vertailtavat tiedot",
-    "sisaanpaasyn-pistemaara": "Sisäänpääsyn pistemäärä"
+    "sisaanpaasyn-pistemaara": "Sisäänpääsyn pistemäärä",
+    "pistemaarasi": "Pistemääräsi",
+    "laske-pistemaarasi": "Laske pistemääräsi"
   }
 }
