+++ src/main/app/src/components/SuosikitVertailuPage/index.tsx
@@ -1,0 +1,166 @@
+import React from 'react';
+
+import { Box, Link, Typography } from '@mui/material';
+import { useTranslation } from 'react-i18next';
+
+import {
+  useSuosikitSelection,
+  useVertailuSuosikit,
+} from '#/src/hooks/useSuosikitSelection';
+import { useTruncatedKuvaus } from '#/src/hooks/useTruncatedKuvaus';
+import { localize } from '#/src/tools/localization';
+
+import { useSuosikitVertailuData } from './useSuosikitVertailuData';
+import { useSuosikitVertailuMask } from './useSuosikitVertailuMask';
+import { VertailuFieldMask } from './VertailuFieldMask';
+import { FIELDS_ORDER } from './vertailuFieldsOrder';
+import { ContentWrapper } from '../common/ContentWrapper';
+import { KorttiLogo } from '../common/KorttiLogo';
+import { MaterialIcon, MaterialIconVariant } from '../common/MaterialIcon';
+import { Murupolku } from '../common/Murupolku';
+import { QueryResult } from '../common/QueryResultWrapper';
+import { TextButton } from '../common/TextButton';
+import { Heading, HeadingBoundary } from '../Heading';
+import { PaperWithTopColor } from '../PaperWithTopColor';
+
+const InfoItem = ({
+  icon,
+  iconVariant = 'filled',
+  label,
+  value,
+}: {
+  icon: MaterialIconName;
+  iconVariant?: MaterialIconVariant;
+  label?: string;
+  value: React.ReactNode;
+}) => {
+  return value ? (
+    <Box display="flex" gap={1}>
+      <MaterialIcon icon={icon} variant={iconVariant} color="primary" />
+      <Box display="flex" flexDirection="column">
+        {label && <Typography>{label}</Typography>}
+        <Typography component="div" display="flex" fontWeight="bold">
+          {value}
+        </Typography>
+      </Box>
+    </Box>
+  ) : null;
+};
+
+const VertailuKortti = ({
+  hakukohde: hakukohde,
+}: {
+  hakukohde: any;
+  removed?: boolean;
+}) => {
+  const { t } = useTranslation();
+  const { toggleVertailu } = useSuosikitSelection();
+
+  const kuvaus = useTruncatedKuvaus(localize(hakukohde.esittely));
+
+  const { mask } = useSuosikitVertailuMask();
+
+  return (
+    <PaperWithTopColor
+      key={hakukohde.hakukohdeOid}
+      role="listitem"
+      sx={{ flex: '1 1 350px' }}>
+      <KorttiLogo
+        entity="oppilaitos"
+        src={hakukohde.logo}
+        alt={t('haku.oppilaitoksen-logo')}
+        display="inline-block"
+        position="relative"
+        maxWidth="50px"
+        marginRight={2}
+        sx={{
+          float: 'left',
+        }}
+      />
+      <Typography variant="body1">{localize(hakukohde.jarjestyspaikka)}</Typography>
+      <Link href={`toteutus/${hakukohde.toteutusOid}`}>
+        <Heading variant="h4" color="primary">
+          {localize(hakukohde.nimi)}
+        </Heading>
+      </Link>
+      <Typography>{kuvaus}</Typography>
+      <Box
+        display="flex"
+        flexDirection="column"
+        justifyContent="space-between"
+        alignItems="stretch"
+        flexWrap="wrap"
+        mt={2}
+        gap={2}>
+        {FIELDS_ORDER.map(({ icon, iconVariant, getLabel, renderValue, fieldId }) => {
+          return mask[fieldId] ? (
+            <InfoItem
+              key={fieldId}
+              icon={icon}
+              iconVariant={iconVariant}
+              label={getLabel?.(t, hakukohde)}
+              value={renderValue?.(hakukohde, t)}
+            />
+          ) : null;
+        })}
+        <hr />
+        <Box display="flex" justifyContent="flex-end" flexWrap="wrap" gap={1}>
+          <TextButton onClick={() => toggleVertailu(hakukohde.hakukohdeOid)}>
+            {t('suosikit.poista-vertailusta')}
+          </TextButton>
+        </Box>
+      </Box>
+    </PaperWithTopColor>
+  );
+};
+
+const Vertailu = ({ oids }: { oids: Array<string> }) => {
+  const queryResult = useSuosikitVertailuData(oids);
+  const { data } = queryResult;
+
+  return (
+    <QueryResult queryResult={queryResult}>
+      <Box display="flex" flexDirection="column" gap={2}>
+        <HeadingBoundary>
+          <Box display="flex" role="list" gap={3} flexWrap="wrap">
+            {data?.map((hakukohdeSuosikki) => (
+              <VertailuKortti
+                key={hakukohdeSuosikki.hakukohdeOid}
+                hakukohde={hakukohdeSuosikki}
+              />
+            ))}
+          </Box>
+        </HeadingBoundary>
+      </Box>
+    </QueryResult>
+  );
+};
+
+export const SuosikitVertailuPage = () => {
+  const { t } = useTranslation();
+  const oids = useVertailuSuosikit();
+
+  const vertailuCount = oids.length;
+
+  return (
+    <ContentWrapper>
+      <Box width="100%" alignSelf="start">
+        <Murupolku
+          path={[
+            { name: t('suosikit.otsikko'), link: 'suosikit' },
+            { name: t('suosikit-vertailu.otsikko') },
+          ]}
+        />
+      </Box>
+      <Heading variant="h1">{t('suosikit-vertailu.otsikko')}</Heading>
+      <HeadingBoundary>
+        <VertailuFieldMask />
+        {vertailuCount > 0 ? (
+          <Vertailu oids={oids} />
+        ) : (
+          t('suosikit-vertailu.ei-vertailtavia-suosikkeja')
+        )}
+      </HeadingBoundary>
+    </ContentWrapper>
+  );
+};
